Cleanup
Some polishing before making the repo public.
