--- .github/workflows/cmake-build-code.yaml
@@ -1,30 +1,0 @@
-name: Build code
-on: [push]
-jobs:
-  run-tests:
-    # Target OS: Ubuntu.
-    runs-on: ubuntu-latest
-    # Use bash for all commands.
-    defaults:
-      run:
-        shell: bash
-    # Set of commands to be run.
-    steps:
-    # Install CMake and other dependencies (if any!).
-    - name: Install dependencies
-      working-directory: ${{github.workspace}}
-      run: sudo apt install cmake
-    # Clone the repository.
-    - uses: actions/checkout@v2
-    # Create the build space.
-    - name: Create build directory
-      working-directory: ${{github.workspace}}
-      run: mkdir ${{github.workspace}}/build
-    # Configure CMake.
-    - name: Invoke CMake
-      working-directory: ${{github.workspace}}/build
-      run: cmake .. -DBUILD_TESTS=ON -DBUILD_DOC=OFF
-    # Build the code (tests included!).
-    - name: Build code
-      working-directory: ${{github.workspace}}/build
-      run: cmake --build .
