+++ doc/extra/main_page.md
@@ -2,6 +2,6 @@ Main Page {#mainpage}
 =========
 
 Sources are available on
-<a href="https://github.com/francofusco/template-cmake-project">
+<a href="https://github.com/francofusco/pendule_pi">
   <img src="GitHub_Logo.png" alt="GitHub" style="height:30px;">
 </a>
