--- include/pendule_pi/__namespace_doc__.hpp
@@ -1,2 +1,0 @@
-/// Main namespace containing (almost) all provided classes.
-namespace pendule_pi { }
